tsn: align name and define convention with other components
Rename configuration components with other components, adopting the
format : CONFIG_RTOS_APPS_<configuration/component>, which already used
by other components such as log and audio.

The convention adopted is the same as Kconfig 3.0: when a configuration
options is enabled, a compile option equal 1 is created, otherwise not
created at all.

Signed-off-by: Tobias Aguiar <tobias.aguiarsilvadealmeida@nxp.com>