types: rename min and max to prevent conflicts with toolchains
Add 'rtos_apps_' prefix to min and max macros to prevent conflicts with
other toolchains, as it is now defined in Zephyr.

Commit 37717b229f51d08121d4bee58931fd03ea57a50c ("sys: util: rename
Z_MIN Z_MAX Z_CLAMP to min max and clamp"), rename Z_MIN to min and
Z_MAX to max, which now creates a redefined warning for min and max
when including 'rtos_apps/types.h'.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>